[refactor]: Carousel module css 자체 style 폴더로 이동
- style 파일이 2개여서 거슬리고 한 번에 컴포넌트 구조가 안보여서 분리함
