+++ component/MainCarousel/CarouselCard.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import styles from './Carousel.module.scss';
+import styles from './styles/Carousel.module.scss';
 import mockImage from '@/assets/img/mockImage.png';
 
 const CarouselCard = () => {
