+++ component/MainCarousel/CarouselContainer.tsx
@@ -3,12 +3,11 @@ import CarouselHeader from './CarouselHeader';
 
 import ImageConstant from './constants';
 import useCarousel from './useCarousel';
-import styles from './Carousel.module.scss';
+import styles from './styles/Carousel.module.scss';
 
 const CarouselContainer = () => {
   const { currentIndex, prev, next, handleTouchStart, handleTouchMove } = useCarousel();
-
-  const mockImgList = new Array(10).fill(null);
+  const mockImageList = new Array(10).fill(null);
 
   return (
     <section>
@@ -19,7 +18,7 @@ const CarouselContainer = () => {
             <div
               className={styles.content}
               style={{ transform: `translateX(-${currentIndex * (100 / ImageConstant.showImage)}%)` }}>
-              {mockImgList.map((_, index) => (
+              {mockImageList.map((_, index) => (
                 <CarouselCard key={index} />
               ))}
             </div>
