+++ component/MainCarousel/styles/Carousel.module.scss
