+++ component/MainCarousel/styles/CarouselHeader.module.scss
